+++ src/tasks/config.dev.js
@@ -3,18 +3,18 @@
 var config = require('./../build.config.js').getConfig();
 var utils = require('./../utils/common.js');
 var cacheBusting = require('./../utils/cacheBusting.js');
+var project = require('./../utils/project.js');
 
-var runKarmaInDev = config.build.spec.runInDev && (utils.hasFiles('src/app', config.app.files.js_spec) || utils.hasFiles('src/common', config.common.files.js_spec));
+var runKarmaInDev = config.build.spec.runInDev && utils.hasFiles(config.app.files.root, config.app.files.js_spec);
+var bowerRc = project.getBowerRc();
 
 var devTasksConfig = {
 
   updateConfig: {
     dev_changeLessSassConfig: {
       update: {
-        'compass.options.outputStyle': 'expanded',
         'compass.options.debugInfo': true,
-        'less.options.dumpLineNumbers': 'mediaquery',
-        'less.options.compress': false
+        'less.options.dumpLineNumbers': 'mediaquery'
       }
     }
   },
@@ -46,10 +46,7 @@ var devTasksConfig = {
         {
           expand: true,
           cwd: '.',
-          src: [
-            '<%= copy.prepare_app_js.options.out %>/**/*.js',
-            '<%= copy.prepare_common_js.options.out %>/**/*.js'
-          ],
+          src: [ '<%= copy.prepare_app_js.options.out %>/**/*.js' ],
           dest: '.'
         }
       ]
@@ -104,33 +101,29 @@ var devTasksConfig = {
      */
     js: {
       files: [
-        config.app.files.js.map(utils.addCwdToPattern('src/app')),
-        config.common.files.js.map(utils.addCwdToPattern('src/common'))
+        config.app.files.js.map(utils.addCwdToPattern(config.app.files.root))
       ],
       tasks: (function () {
         return [].concat(
           utils.includeIf('jshint:src', config.build.jshint.runInDev),
           utils.includeIf('karma:dev_spec:run', runKarmaInDev),
           'copy:prepare_app_js',
-          'copy:prepare_common_js',
           'replace:dev_cacheBusting',
-          'indexHtml:prepare'
+          'processHtml:prepare'
         );
       })()
     },
 
     js_mock: {
       files: [
-        config.app.files.js_mock.map(utils.addCwdToPattern('src/app')),
-        config.common.files.js_mock.map(utils.addCwdToPattern('src/common'))
+        config.app.files.js_mock.map(utils.addCwdToPattern(config.app.files.root))
       ],
       tasks: (function () {
         return [].concat(
           utils.includeIf('jshint:mock', config.build.jshint.runInDev),
           utils.includeIf('karma:dev_spec:run', runKarmaInDev),
           utils.includeIf('copy:prepare_app_js_mock', config.build.mocks.loadInBrowser),
-          utils.includeIf('copy:prepare_common_js_mock', config.build.mocks.loadInBrowser),
-          'indexHtml:prepare'
+          'processHtml:prepare'
         );
       })()
     },
@@ -144,8 +137,7 @@ var devTasksConfig = {
         livereload: false
       },
       files: [
-        config.app.files.js_spec.map(utils.addCwdToPattern('src/app')),
-        config.common.files.js_spec.map(utils.addCwdToPattern('src/common'))
+        config.app.files.js_spec.map(utils.addCwdToPattern(config.app.files.root))
       ],
       tasks: (function () {
         return [].concat(
@@ -160,8 +152,7 @@ var devTasksConfig = {
         livereload: false
       },
       files: [
-        config.app.files.js_e2e.map(utils.addCwdToPattern('src/app')),
-        config.common.files.js_e2e.map(utils.addCwdToPattern('src/common'))
+        config.app.files.js_e2e.map(utils.addCwdToPattern(config.app.files.root))
       ],
       tasks: (function () {
         return [].concat(
@@ -172,7 +163,7 @@ var devTasksConfig = {
 
     translations: {
       files: [
-        config.app.files.translations.map(utils.addCwdToPattern('src/app'))
+        config.app.files.translations.map(utils.addCwdToPattern(config.app.files.root))
       ],
       tasks: [ 'translations2js:prepare' ]
     },
@@ -189,25 +180,15 @@ var devTasksConfig = {
     },
 
     html: {
-      files: [ 'src/index.html' ],
-      tasks: (function () {
-        return [].concat(
-          'indexHtml:prepare'
-        );
-      })()
+      files: config.app.files.html.map(utils.addCwdToPattern(config.app.files.root)),
+      tasks: [ 'processHtml:prepare' ]
     },
 
     templates: {
       files: [
-        config.app.files.templates.map(utils.addCwdToPattern('src/app')),
-        config.common.files.templates.map(utils.addCwdToPattern('src/common'))
+        config.app.files.templates.map(utils.addCwdToPattern(config.app.files.root))
       ],
-      tasks: (function () {
-        return [].concat(
-          'copy:prepare_app_templates',
-          'copy:prepare_common_templates'
-        );
-      })()
+      tasks: [ 'copy:prepare_app_templates' ]
     },
 
     less: {
@@ -219,10 +200,9 @@ var devTasksConfig = {
       ],
       tasks: (function () {
         return [].concat(
-          utils.includeIf('less:prepare_app', config.build.less.enabled && utils.hasFiles('src/app', config.app.files.less)),
-          utils.includeIf('less:prepare_common', config.build.less.enabled && utils.hasFiles('src/common', config.common.files.less)),
+          utils.includeIf('less:prepare_app', config.build.less.enabled && utils.hasFiles(config.app.files.root, config.app.files.less)),
           'concat:prepare_css',
-          'indexHtml:prepare'
+          'processHtml:prepare'
         );
       })()
     },
@@ -235,10 +215,9 @@ var devTasksConfig = {
       ],
       tasks: (function () {
         return [].concat(
-          utils.includeIf('compass:prepare_app', config.build.sass.enabled && utils.hasFiles('src/app', config.app.files.sass)),
-          utils.includeIf('compass:prepare_common', config.build.sass.enabled && utils.hasFiles('src/common', config.common.files.sass)),
+          utils.includeIf('compass:prepare_app', config.build.sass.enabled && utils.hasFiles(config.app.files.root, config.app.files.sass)),
           'concat:prepare_css',
-          'indexHtml:prepare'
+          'processHtml:prepare'
         );
       })()
     },
@@ -260,7 +239,16 @@ var devTasksConfig = {
       ]
     }
   }
-
 };
+
+if (config.build.bower.runInDev) {
+  devTasksConfig.watch.vendor = {
+    options: {
+      livereload: false
+    },
+    files: [ bowerRc.json ],
+    tasks: ['shell:bower']
+  };
+}
 
 module.exports = devTasksConfig;